Split parser into modules
The TimeType::parse() function was removed because we don't want that
there. It is not that nice, because the interface would return either a
TimeType or an Iterator.
